move add_parametric_commuting_paulis_exp_gate function to core.circuit
