+++ packages/openfermion/tests/openfermion/ansatz/test_uccsd.py
@@ -15,6 +15,7 @@ from openfermion import FermionOperator
 
 from quri_parts.chem.utils.excitations import excitations
 from quri_parts.circuit import LinearMappedUnboundParametricQuantumCircuit
+from quri_parts.core.circuit import add_parametric_commuting_paulis_exp_gate
 from quri_parts.core.operator import Operator, pauli_label, truncate
 from quri_parts.openfermion.ansatz.uccsd import (
     TrotterUCCSD,
@@ -27,10 +28,7 @@ from quri_parts.openfermion.transforms import (
     jordan_wigner,
     symmetry_conserving_bravyi_kitaev,
 )
-from quri_parts.openfermion.utils import (
-    add_exp_excitation_gates_trotter_decomposition,
-    add_parametric_commuting_paulis_exp_gate,
-)
+from quri_parts.openfermion.utils import add_exp_excitation_gates_trotter_decomposition
 from quri_parts.openfermion.utils.add_exp_excitation_gates_trotter_decomposition import (  # noqa: E501
     create_anti_hermitian_sd_excitation_operator,
 )
